requirement.txt
fastapi
uvicorn
kafka-python
tweepy
praw
sqlalchemy
psycopg2-binary
pydantic
redis
pandas
scikit-learn
transformers
torch
praw
tweepy
pyspark
pydantic_settings